# docs/swagger.yaml
basePath: /
definitions:
  models.Nas:
    properties:
      Community:
        type: string
      Description:
        type: string
      Ports:
        type: integer
      Secret:
        type: string
      Server:
        type: string
      Shortname:
        type: string
      Type:
        type: string
      Username:
        type: string
      id:
        type: integer
    type: object
  models.Radcheck:
    properties:
      Attribute:
        type: string
      Groups:
        items:
          $ref: '#/definitions/models.Radusergroup'
        type: array
      Options:
        items:
          $ref: '#/definitions/models.Radreply'
        type: array
      Username:
        type: string
      Value:
        type: string
      id:
        type: integer
    type: object
  models.Radgroupreply:
    properties:
      Attribute:
        type: string
      Groupname:
        type: string
      Value:
        type: string
      id:
        type: integer
    type: object
  models.Radreply:
    properties:
      Attribute:
        type: string
      Value:
        type: string
      id:
        type: integer
    type: object
  models.Radusergroup:
    properties:
      Groupname:
        type: string
      Priority:
        type: integer
      id:
        type: integer
    type: object
info:
  contact: {}
  description: API for managing FreeRADIUS users and NAS devices
  title: Radius API
  version: "1.0"
paths:
  /api/accounting/{username}:
    get:
      consumes:
      - application/json
      description: Get a specific accounting record
      parameters:
      - description: Username of the accounting record
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data of the accounting record
        "500":
          description: Error
      summary: Get a specific accounting record
      tags:
      - Accounting
  /api/groups:
    get:
      consumes:
      - application/json
      description: Get all Group
      produces:
      - application/json
      responses:
        "200":
          description: List of all Group
        "500":
          description: Error
      summary: Get all Group
      tags:
      - Group
    post:
      consumes:
      - application/json
      description: Add Group
      parameters:
      - description: Group data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Radgroupreply'
      produces:
      - application/json
      responses:
        "200":
          description: Group added successfully
        "500":
          description: Error Group not added successfully
      summary: Add Group
      tags:
      - Group
  /api/groups/{groupname}:
    delete:
      description: Delete Group
      parameters:
      - description: Groupname of the group to delete
        in: path
        name: groupname
        required: true
        type: string
      responses:
        "200":
          description: Group deleted successfully
        "500":
          description: Error
      summary: Delete Group
      tags:
      - Group
    get:
      consumes:
      - application/json
      description: Get a specific Group
      parameters:
      - description: Groupname of the Group
        in: path
        name: groupname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data of the Group
        "500":
          description: Error
      summary: Get a specific Group
      tags:
      - Group
    patch:
      consumes:
      - application/json
      description: Update a Group
      parameters:
      - description: Groupname of the Group to update
        in: path
        name: groupname
        required: true
        type: string
      - description: Data to update the Group
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Radgroupreply'
      produces:
      - application/json
      responses:
        "200":
          description: Group updated successfully
        "500":
          description: Error
      summary: Update a Group
      tags:
      - Group
  /api/nas:
    get:
      consumes:
      - application/json
      description: Get all NAS
      produces:
      - application/json
      responses:
        "200":
          description: List of all NAS
        "500":
          description: Error
      summary: Get all NAS
      tags:
      - NAS
    post:
      consumes:
      - application/json
      description: Add NAS
      parameters:
      - description: NAS data
        in: body
        name: nas
        required: true
        schema:
          $ref: '#/definitions/models.Nas'
      produces:
      - application/json
      responses:
        "200":
          description: NAS added successfully
        "500":
          description: Error NAS not added successfully
      summary: Add NAS
      tags:
      - NAS
  /api/nas/{nasname}:
    delete:
      description: Delete NAS
      parameters:
      - description: Nasname of the NAS to delete
        in: path
        name: nasname
        required: true
        type: string
      responses:
        "200":
          description: Nas deleted successfully
        "500":
          description: Error
      summary: Delete NAS
      tags:
      - NAS
    get:
      consumes:
      - application/json
      description: Get a specific NAS
      parameters:
      - description: Nasname of the NAS
        in: path
        name: nasname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data of the NAS
        "500":
          description: Error
      summary: Get a specific NAS
      tags:
      - NAS
    patch:
      consumes:
      - application/json
      description: Update a NAS
      parameters:
      - description: Nasname of the NAS to update
        in: path
        name: nasname
        required: true
        type: string
      - description: Data to update the NAS
        in: body
        name: nas
        required: true
        schema:
          $ref: '#/definitions/models.Nas'
      produces:
      - application/json
      responses:
        "200":
          description: Nas updated successfully
        "500":
          description: Error
      summary: Update a NAS
      tags:
      - NAS
  /api/users:
    get:
      consumes:
      - application/json
      description: Get all radius users
      produces:
      - application/json
      responses:
        "200":
          description: List of all user
        "500":
          description: Error
      summary: Get all radius users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new Radius user with radcheck and associated radreply
        options.
      parameters:
      - description: New Radius user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Radcheck'
      produces:
      - application/json
      responses:
        "200":
          description: User added successfully
        "500":
          description: Error
      summary: Create a new Radius user
      tags:
      - users
  /api/users/{username}:
    delete:
      consumes:
      - application/json
      description: Delete radius user
      parameters:
      - description: Username of radius user
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
        "500":
          description: Error
      summary: Delete radius user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get specific radius user
      parameters:
      - description: Username of radius user
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "500":
          description: Error
      summary: Get specific radius user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update radius user
      parameters:
      - description: Username of radius user
        in: path
        name: username
        required: true
        type: string
      - description: Radius user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Radcheck'
      produces:
      - application/json
      responses:
        "200":
          description: User uptaded successfully
        "500":
          description: Error
      summary: Update radius user
      tags:
      - users
  /api/users/{username}/options/{optionname}:
    delete:
      consumes:
      - application/json
      description: Delete option of radius user
      parameters:
      - description: Username of radius user
        in: path
        name: username
        required: true
        type: string
      - description: Radius user option data
        in: body
        name: option
        required: true
        schema:
          $ref: '#/definitions/models.Radreply'
      produces:
      - application/json
      responses:
        "200":
          description: Option user deleted successfully
        "500":
          description: Error
      summary: Delete option of radius user
      tags:
      - users
swagger: "2.0"